Fixed Bug in Partial Tree Iteration
+ In cases where the selected starting node did not have a sibling, the
ending node was set to nullptr. Normally, the nullptr is reserved to
indicate the end of the tree, and the partial tree traversals would
therefore continue until the end of the tree was reached. This has now
been fixed by properly computing the node that occurs immediately after
the last node in the iteration sequence. This bug affected all iterator
types.
+ Added unit tests to cover these edge cases.